requirements.txt
gftools @ git+https://github.com/m4rc1e/tools@trafficjam2
fontbakery[googlefonts]
fonttools
gflanguages
PyGithub
tqdm
Jinja2
anybadge
humanize
github3api
